#44 refactor: reorganize knowledge model package structure
