+++ src/main/java/com/kairos/context_engine/domain/model/knowledge/KnowledgeTriple.java
@@ -1,4 +1,6 @@
-package com.kairos.context_engine.domain.model;
+package com.kairos.context_engine.domain.model.knowledge;
+
+import com.kairos.context_engine.domain.model.Triple;
 
 import java.util.UUID;
 
@@ -6,23 +8,26 @@ public record KnowledgeTriple(
         Concept subject,
         String predicate,
         Concept object,
-        UUID chunkId
+        UUID chunkId,
+        double weight
 ) {
     public static KnowledgeTriple create(Triple triple, UUID chunkId) {
         return new KnowledgeTriple(
                 Concept.create(triple.subject()),
                 triple.predicate(),
                 Concept.create(triple.object()),
-                chunkId
+                chunkId,
+                triple.weight()
         );
     }
 
-    public static KnowledgeTriple create(String subject, String predicate, String object, UUID chunkId) {
+    public static KnowledgeTriple create(String subject, String predicate, String object, UUID chunkId, double weight) {
         return new KnowledgeTriple(
                 Concept.create(subject),
                 predicate,
                 Concept.create(object),
-                chunkId
+                chunkId,
+                weight
         );
     }
 }
