+++ src/main/java/com/kairos/context_engine/presentation/dto/response/ContextResponse.java
@@ -1,7 +1,7 @@
 package com.kairos.context_engine.presentation.dto.response;
 
 import com.kairos.context_engine.domain.model.content.Chunk;
-import com.kairos.context_engine.domain.model.KnowledgeTriple;
+import com.kairos.context_engine.domain.model.knowledge.KnowledgeTriple;
 
 import java.util.List;
