+++ src/main/java/com/kairos/context_engine/presentation/dto/response/KnowledgeTripleResponse.java
@@ -1,6 +1,6 @@
 package com.kairos.context_engine.presentation.dto.response;
 
-import com.kairos.context_engine.domain.model.KnowledgeTriple;
+import com.kairos.context_engine.domain.model.knowledge.KnowledgeTriple;
 
 import java.util.UUID;
